requirements.txt
streamlit>=1.33.0
polars>=0.20.0
psycopg2-binary>=2.9.9
plotly>=5.22.0
pandas>=2.2.2
